Replaced data input in deterministic analysis
This push includes an important correction to the interpretation of the AC_LOF vs AC_NO_LOF subpopulations. The AC_LOF arm, which only appears in the standard care route, should actually include both LOF and NO_LOF individuals. The basic formatting of the code is such that the AC arm is divided into LOF and NO_LOF subgroups, but we need to restructure the simulation to reflect that the values in the data represent a combined cohort of both genotypes.